move AnalogFS
