+++ src/iptux/RecvFileData.cpp
@@ -19,7 +19,7 @@
 #include <sys/socket.h>
 
 #include "iptux-core/Command.h"
-#include "iptux/AnalogFS.h"
+#include "iptux-core/AnalogFS.h"
 #include "iptux-core/deplib.h"
 #include "iptux/global.h"
 #include "iptux-core/ipmsg.h"
